Always update the branch status in status bar
Originally when adding files the 'index:' and 'working:' values in the status
bar would be updated reflecting the current branch status. However when using
reset the changes to the branch would not be immediately reflected in the
status bar. The branch status would then be updated in the status bar after
saving the file. This change makes it so that after any git command is run the
branch status is updated in the status bar.