Add option to show untracked files on the diff output
This stages the untracked files using the intent-to-add flag, -N.
Essentially, a 0-byte version of the files is staged, allowing the full file to be show as a difference.
This operation is reversed as soon as the diff output is generated by reseting the files back to their untracked status.